+++ example/a.ts
@@ -1,11 +1,14 @@
-import { Elysia } from '../src'
+import { Elysia, t } from '../src'
+import { z } from 'zod'
+import { req } from '../test/utils'
 
-// This uses aot: true by default in 1.4 (broken on Bun)
-const app = new Elysia({ systemRouter: true })
-  .get("/", "Hello Elysia")
-  .get("/json", () => ({ message: "Hello World", timestamp: Date.now() }))
-
-Bun.serve({
-	port: 3000,
-	fetch: app.fetch
-})
+export const app = new Elysia()
+	.ws('/', {
+		open(ws) {
+			ws.subscribe('a')
+		},
+		message(a) {
+			console.log(a.subscriptions)
+		}
+	})
+	.listen(3000)